Consolidate derived media metadata
