+++ mobile/apps/photos/lib/module/metadata/exif.dart
@@ -1,8 +1,12 @@
 import "dart:async";
 import "dart:io";
+import 'dart:typed_data';
 
 import "package:computer/computer.dart";
 import 'package:exif_reader/exif_reader.dart';
+import 'package:image/image.dart' as img;
+// ignore: implementation_imports
+import 'package:image/src/util/rational.dart' as img_util;
 import 'package:intl/intl.dart';
 import 'package:logging/logging.dart';
 import 'package:photos/models/file/file.dart';
@@ -24,6 +28,96 @@ const _exifDateTimePattern = "yyyy:MM:dd HH:mm:ss";
 const _emptyExifDateTime = "0000:00:00 00:00:00";
 
 final _logger = Logger("ExifUtil");
+
+// dart format off
+const _copiedExifFields = [
+  (ifd: "Image", src: "Make",                    dst: "Make",                    render: false),
+  (ifd: "Image", src: "Model",                   dst: "Model",                   render: false),
+  (ifd: "Image", src: "Orientation",             dst: "Orientation",             render: true),
+  (ifd: "Image", src: "Artist",                  dst: "Artist",                  render: false),
+  (ifd: "Image", src: "Copyright",               dst: "Copyright",               render: false),
+  (ifd: "EXIF",  src: "DateTimeOriginal",        dst: "DateTimeOriginal",        render: false),
+  (ifd: "EXIF",  src: "DateTimeDigitized",       dst: "DateTimeDigitized",       render: false),
+  (ifd: "EXIF",  src: "OffsetTimeOriginal",      dst: "OffsetTimeOriginal",      render: false),
+  (ifd: "EXIF",  src: "LensModel",               dst: "LensModel",               render: false),
+  (ifd: "EXIF",  src: "ExposureTime",            dst: "ExposureTime",            render: false),
+  (ifd: "EXIF",  src: "FNumber",                 dst: "FNumber",                 render: false),
+  (ifd: "EXIF",  src: "ISOSpeedRatings",         dst: "ISOSpeed",                render: false),
+  (ifd: "EXIF",  src: "FocalLength",             dst: "FocalLength",             render: false),
+  (ifd: "EXIF",  src: "FocalLengthIn35mmFilm",   dst: "FocalLengthIn35mmFilm",   render: false),
+  (ifd: "EXIF",  src: "ColorSpace",              dst: "ColorSpace",              render: true),
+  (ifd: "GPS",   src: "GPSLatitudeRef",          dst: "GPSLatitudeRef",          render: false),
+  (ifd: "GPS",   src: "GPSLatitude",             dst: "GPSLatitude",             render: false),
+  (ifd: "GPS",   src: "GPSLongitudeRef",         dst: "GPSLongitudeRef",         render: false),
+  (ifd: "GPS",   src: "GPSLongitude",            dst: "GPSLongitude",            render: false),
+  (ifd: "GPS",   src: "GPSAltitudeRef",           dst: "GPSAltitudeRef",          render: false),
+  (ifd: "GPS",   src: "GPSAltitude",              dst: "GPSAltitude",             render: false),
+  (ifd: "GPS",   src: "GPSMapDatum",              dst: "GPSMapDatum",             render: false),
+];
+// dart format on
+
+img.IfdValue? _imageExifValue(IfdTag? tag) {
+  final values = tag?.values;
+  if (tag == null || values == null || values is IfdNone) {
+    return null;
+  }
+
+  List<int> ints() => values.toList().cast<int>();
+  List<double> doubles() =>
+      values.toList().map((value) => (value as num).toDouble()).toList();
+  List<img_util.Rational> ratios() => values
+      .toList()
+      .cast<Ratio>()
+      .map((value) => img_util.Rational(value.numerator, value.denominator))
+      .toList();
+
+  return switch (tag.tagType) {
+    "Byte" => img.IfdByteValue.list(Uint8List.fromList(ints())),
+    "ASCII" => img.IfdValueAscii(
+      String.fromCharCodes(ints().takeWhile((byte) => byte != 0)),
+    ),
+    "Short" => img.IfdValueShort.list(ints()),
+    "Long" || "IFD" => img.IfdValueLong.list(ints()),
+    "Signed Byte" => img.IfdValueSByte.list(ints()),
+    "Undefined" => img.IfdValueUndefined.list(ints()),
+    "Signed Short" => img.IfdValueSShort.list(ints()),
+    "Signed Long" => img.IfdValueSLong.list(ints()),
+    "Ratio" => img.IfdValueRational.list(ratios()),
+    "Signed Ratio" => img.IfdValueSRational.list(ratios()),
+    "Single-Precision Floating Point (32-bit)" => img.IfdValueSingle.list(
+      doubles(),
+    ),
+    "Double-Precision Floating Point (64-bit)" => img.IfdValueDouble.list(
+      doubles(),
+    ),
+    _ => null,
+  };
+}
+
+// TODO: Move EXIF writes off the Dart image library to a Rust implementation.
+Future<void> copyExif(
+  EnteFile source,
+  img.Image destination, {
+  bool copyRenderingFields = true,
+}) async {
+  final sourceExif = await getExif(source);
+  for (final field in _copiedExifFields) {
+    if (!copyRenderingFields && field.render) {
+      continue;
+    }
+    final destinationIfd = switch (field.ifd) {
+      "Image" => destination.exif.imageIfd,
+      "EXIF" => destination.exif.exifIfd,
+      "GPS" => destination.exif.gpsIfd,
+      _ => throw UnsupportedError("Unknown EXIF IFD: ${field.ifd}"),
+    };
+    final value = _imageExifValue(sourceExif["${field.ifd} ${field.src}"]);
+    if (value != null) {
+      destinationIfd[field.dst] = value;
+    }
+  }
+}
+
 final _standardExifDateTimePattern = RegExp(
   r'^(\d{4}:(0[1-9]|1[0-2]):(0[1-9]|[12]\d|3[01]) ([01]\d|2[0-3]):([0-5]\d):([0-5]\d))([\.:]\d+)?$',
 );
