+++ mobile/apps/photos/lib/ui/actions/file/file_actions.dart
@@ -8,13 +8,13 @@ import "package:photos/events/details_sheet_event.dart";
 import "package:photos/generated/l10n.dart";
 import "package:photos/models/file/extensions/file_props.dart";
 import 'package:photos/models/file/file.dart';
+import 'package:photos/module/metadata/panorama.dart';
 import "package:photos/service_locator.dart";
 import "package:photos/services/media_store_service.dart";
 import "package:photos/theme/ente_theme.dart";
 import "package:photos/ui/notification/toast.dart";
 import 'package:photos/ui/viewer/file/file_details_widget.dart';
 import "package:photos/utils/delete_file_util.dart";
-import "package:photos/utils/panorama_util.dart";
 
 Future<void> showSingleFileDeleteSheet(
   BuildContext context,
